Unify AudioWidget and VideoWidget using MediaWidget
